refactor integration test helpers
