+++ internal/integration/export_import_test.go
@@ -2,10 +2,7 @@ package integration_test
 
 import (
 	"bytes"
-	"compress/gzip"
 	"context"
-	"database/sql"
-	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
@@ -16,65 +13,16 @@ import (
 	"sort"
 	"testing"
 
-	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/pgxpool"
-	"github.com/riverqueue/river"
 	"github.com/testcontainers/testcontainers-go"
 	tcminio "github.com/testcontainers/testcontainers-go/modules/minio"
 	"github.com/testcontainers/testcontainers-go/wait"
 
 	"era/booru/internal/config"
-	"era/booru/internal/db"
-	"era/booru/internal/minio"
-	"era/booru/internal/processing"
-	"era/booru/internal/queue"
-	qworkers "era/booru/internal/queue/workers"
+	common "era/booru/internal/integration/common"
 	"era/booru/internal/server"
 
 	"time"
-
-	mc "github.com/minio/minio-go/v7"
 )
-
-func parseExport(data []byte) ([]map[string]any, error) {
-	gz, err := gzip.NewReader(bytes.NewReader(data))
-	if err != nil {
-		return nil, err
-	}
-	dec := json.NewDecoder(gz)
-	var meta map[string]any
-	if err := dec.Decode(&meta); err != nil {
-		return nil, err
-	}
-	var items []map[string]any
-	for {
-		var item map[string]any
-		if err := dec.Decode(&item); err != nil {
-			if err == io.EOF {
-				break
-			}
-			return nil, err
-		}
-		items = append(items, item)
-	}
-	return items, nil
-}
-
-func waitForPostgres(dsn string, timeout time.Duration) error {
-	deadline := time.Now().Add(timeout)
-	for time.Now().Before(deadline) {
-		db, err := sql.Open("postgres", dsn)
-		if err == nil {
-			if err = db.Ping(); err == nil {
-				db.Close()
-				return nil
-			}
-			db.Close()
-		}
-		time.Sleep(200 * time.Millisecond)
-	}
-	return fmt.Errorf("Postgres not ready after %s", timeout)
-}
 
 func TestExportImportCycle(t *testing.T) {
 	if os.Getenv("RUN_INTEGRATION_TESTS") == "" {
@@ -120,7 +68,7 @@ func TestExportImportCycle(t *testing.T) {
 	}
 	defer mC.Terminate(ctx)
 
-	waitForPostgres(dsn, 30*time.Second)
+	common.WaitForPostgres(dsn, 30*time.Second)
 
 	minioAddr, err := mC.ConnectionString(ctx)
 	if err != nil {
@@ -152,7 +100,7 @@ func TestExportImportCycle(t *testing.T) {
 	defer srv.Close()
 
 	// Start the media worker
-	mediaWorker, err := startMediaWorker(ctx, cfg)
+	mediaWorker, err := common.StartMediaWorker(ctx, cfg)
 	if err != nil {
 		t.Fatalf("media worker: %v", err)
 	}
@@ -162,93 +110,35 @@ func TestExportImportCycle(t *testing.T) {
 	defer ts.Close()
 	client := ts.Client()
 
-	waitFor := func(id string) {
-		deadline := time.Now().Add(10 * time.Second)
-		for time.Now().Before(deadline) {
-			resp, err := client.Get(ts.URL + "/api/media/" + id)
-			if err == nil && resp.StatusCode == http.StatusOK {
-				resp.Body.Close()
-				return
-			}
-			if resp != nil {
-				resp.Body.Close()
-			}
-			time.Sleep(600 * time.Millisecond)
-		}
-		t.Fatalf("media %s not ingested", id)
+	img1Hash, err := common.UploadAndWait(ctx, srv.Minio, client, ts.URL, filepath.Join("testdata", "img1.png"))
+	if err != nil {
+		t.Fatalf("upload img1: %v", err)
+	}
+	img2Hash, err := common.UploadAndWait(ctx, srv.Minio, client, ts.URL, filepath.Join("testdata", "img2.png"))
+	if err != nil {
+		t.Fatalf("upload img2: %v", err)
+	}
+	img3Hash, err := common.UploadAndWait(ctx, srv.Minio, client, ts.URL, filepath.Join("testdata", "img3.png"))
+	if err != nil {
+		t.Fatalf("upload img3: %v", err)
 	}
 
-	upload := func(obj, path string) string {
-		f, err := os.Open(path)
-		if err != nil {
-			t.Fatalf("open %s: %v", path, err)
-		}
-		defer f.Close()
-
-		// Compute hash of the file
-		hash, err := processing.HashFile128Hex(f)
-		if err != nil {
-			t.Fatalf("hash %s: %v", path, err)
-		}
-
-		// Reset file pointer to beginning
-		if _, err := f.Seek(0, 0); err != nil {
-			t.Fatalf("seek %s: %v", path, err)
-		}
-
-		// Upload with hash as object name (no extension)
-		if _, err := srv.Minio.PutObject(ctx, srv.Minio.Bucket, hash, f, -1, mc.PutObjectOptions{ContentType: "image/png"}); err != nil {
-			t.Fatalf("put %s: %v", hash, err)
-		}
-
-		waitFor(hash)
-		return hash
+	if err := common.AddTags(client, ts.URL, img1Hash, []string{"alpha"}); err != nil {
+		t.Fatalf("tags: %v", err)
+	}
+	if err := common.AddTags(client, ts.URL, img2Hash, []string{"beta"}); err != nil {
+		t.Fatalf("tags: %v", err)
+	}
+	if err := common.AddTags(client, ts.URL, img3Hash, []string{"gamma"}); err != nil {
+		t.Fatalf("tags: %v", err)
 	}
 
-	img1Hash := upload("img1.png", filepath.Join("testdata", "img1.png"))
-	img2Hash := upload("img2.png", filepath.Join("testdata", "img2.png"))
-	img3Hash := upload("img3.png", filepath.Join("testdata", "img3.png"))
-
-	addTags := func(id string, tags []string) {
-		b, _ := json.Marshal(struct {
-			Tags []string `json:"tags"`
-		}{Tags: tags})
-		resp, err := client.Post(ts.URL+"/api/media/"+id+"/tags", "application/json", bytes.NewReader(b))
-		if err != nil {
-			t.Fatalf("post tags %s: %v", id, err)
-		}
-		resp.Body.Close()
-		if resp.StatusCode != http.StatusOK {
-			t.Fatalf("tag response %d", resp.StatusCode)
-		}
+	if err := common.SetUploadDate(client, ts.URL, img1Hash, "2021-01-02"); err != nil {
+		t.Fatalf("set date: %v", err)
 	}
-
-	addTags(img1Hash, []string{"alpha"})
-	addTags(img2Hash, []string{"beta"})
-	addTags(img3Hash, []string{"gamma"})
-
-	setDate := func(id, val string) {
-		b, _ := json.Marshal(struct {
-			Dates []struct {
-				Name  string `json:"name"`
-				Value string `json:"value"`
-			} `json:"dates"`
-		}{Dates: []struct {
-			Name  string `json:"name"`
-			Value string `json:"value"`
-		}{{Name: "upload", Value: val}}})
-		resp, err := client.Post(ts.URL+"/api/media/"+id+"/dates", "application/json", bytes.NewReader(b))
-		if err != nil {
-			t.Fatalf("post date %s: %v", id, err)
-		}
-		resp.Body.Close()
-		if resp.StatusCode != http.StatusOK {
-			t.Fatalf("date response %d", resp.StatusCode)
-		}
+	if err := common.SetUploadDate(client, ts.URL, img2Hash, "2022-02-03"); err != nil {
+		t.Fatalf("set date: %v", err)
 	}
-
-	setDate(img1Hash, "2021-01-02")
-	setDate(img2Hash, "2022-02-03")
 
 	resp, err := client.Get(ts.URL + "/api/admin/export-tags")
 	if err != nil {
@@ -259,7 +149,7 @@ func TestExportImportCycle(t *testing.T) {
 	if resp.StatusCode != http.StatusOK {
 		t.Fatalf("export response %d", resp.StatusCode)
 	}
-	items1, err := parseExport(first)
+	items1, err := common.ParseExport(first)
 	if err != nil {
 		t.Fatalf("parse export: %v", err)
 	}
@@ -285,7 +175,7 @@ func TestExportImportCycle(t *testing.T) {
 	}
 	defer srv.Close()
 
-	mediaWorker, err = startMediaWorker(ctx, cfg)
+	mediaWorker, err = common.StartMediaWorker(ctx, cfg)
 	if err != nil {
 		t.Fatalf("restart media worker: %v", err)
 	}
@@ -306,9 +196,15 @@ func TestExportImportCycle(t *testing.T) {
 		t.Fatalf("regenerate response %d", resp.StatusCode)
 	}
 
-	waitFor(img1Hash)
-	waitFor(img2Hash)
-	waitFor(img3Hash)
+	if err := common.WaitForMedia(client, ts.URL, img1Hash, 10*time.Second); err != nil {
+		t.Fatalf("wait img1: %v", err)
+	}
+	if err := common.WaitForMedia(client, ts.URL, img2Hash, 10*time.Second); err != nil {
+		t.Fatalf("wait img2: %v", err)
+	}
+	if err := common.WaitForMedia(client, ts.URL, img3Hash, 10*time.Second); err != nil {
+		t.Fatalf("wait img3: %v", err)
+	}
 
 	resp, err = client.Post(ts.URL+"/api/admin/import-tags", "application/gzip", bytes.NewReader(first))
 	if err != nil {
@@ -330,7 +226,7 @@ func TestExportImportCycle(t *testing.T) {
 	if resp.StatusCode != http.StatusOK {
 		t.Fatalf("re-export response %d", resp.StatusCode)
 	}
-	items2, err := parseExport(second)
+	items2, err := common.ParseExport(second)
 	if err != nil {
 		t.Fatalf("parse second export: %v", err)
 	}
@@ -340,65 +236,5 @@ func TestExportImportCycle(t *testing.T) {
 
 	if !reflect.DeepEqual(items1, items2) {
 		t.Fatalf("export/import mismatch")
-	}
-}
-
-func startMediaWorker(ctx context.Context, cfg *config.Config) (*MediaWorkerWrapper, error) {
-	pool, err := pgxpool.New(ctx, cfg.PostgresDSN)
-	if err != nil {
-		return nil, err
-	}
-
-	workers := river.NewWorkers()
-	client, err := queue.NewClient(ctx, pool, workers, queue.ClientTypeMediaWorker)
-	if err != nil {
-		return nil, err
-	}
-
-	database, err := db.New(cfg, client)
-	if err != nil {
-		return nil, err
-	}
-
-	m, err := minio.New(cfg)
-	if err != nil {
-		return nil, err
-	}
-
-	// Register BOTH workers (even though media worker only processes ProcessWorker)
-	river.AddWorker(workers, &qworkers.ProcessWorker{
-		Minio: m,
-		DB:    database,
-		Cfg:   cfg,
-	})
-
-	// Add this - needed for media worker to enqueue index jobs
-	river.AddWorker(workers, &qworkers.IndexWorker{
-		DB: database,
-	})
-
-	if err := client.Start(ctx); err != nil {
-		return nil, err
-	}
-
-	return &MediaWorkerWrapper{
-		client: client,
-		pool:   pool,
-		ctx:    ctx,
-	}, nil
-}
-
-type MediaWorkerWrapper struct {
-	client *river.Client[pgx.Tx]
-	pool   *pgxpool.Pool
-	ctx    context.Context
-}
-
-func (w *MediaWorkerWrapper) Stop() {
-	if w.client != nil {
-		w.client.Stop(context.Background())
-	}
-	if w.pool != nil {
-		w.pool.Close()
 	}
 }
